% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexml_validate.R
\name{nexml_validate}
\alias{nexml_validate}
\title{validate nexml using the online validator tool}
\usage{
nexml_validate(file, schema = CANONICAL_SCHEMA)
}
\arguments{
\item{file}{path to the nexml file to validate}

\item{schema}{URL of schema (for fallback method only, set by default).}
}
\value{
TRUE if the file is valid, FALSE or error message otherwise
}
\description{
validate nexml using the online validator tool
}
\details{
Requires an internet connection.  see http://www.nexml.org/nexml/phylows/validator for more information in debugging invalid files
}
\examples{
\dontrun{
f <- system.file("extdata/example.xml", package="nexld")
nexml_validate("birds_orders.xml")
}
}
