% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexld.R
\name{xml_to_json}
\alias{xml_to_json}
\title{xml_to_json}
\usage{
xml_to_json(x, out = NULL)
}
\arguments{
\item{x}{path to a nexml file}
}
\value{
json string
}
\description{
Convert an NeXML file into a json string
}
\examples{

ex <- system.file("extdata/example.xml", package = "nexld")
xml_to_json(ex)

}
